+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 =========
+## Unreleased
+### Changed
+- Reset query relations table joins on new execution of query.
 
 ## 1.0.0-rc.24 - 2018-07-26
 ### Fixed